change type data migration

+++ db/migrations/20241204062956_create_table_shippings.up.sql
@@ -1,11 +1,10 @@
 CREATE TABLE "shippings" (
   "id" bigserial PRIMARY KEY,
   "name" varchar NOT NULL,
-  "price" decimal NOT NULL,
+  "price" float NOT NULL,
   "state" varchar NOT NULL,
   "updated_at" timestamptz NOT NULL DEFAULT (now()),
   "created_at" timestamptz NOT NULL DEFAULT (now())
 );
-
 
 COMMENT ON COLUMN "shippings"."price" IS 'must be positive';
